Updates to `snarkvm-parameters` and expands serial numbers tree and commitments tree to depth 42

+++ dpc/src/ledger/commitments.rs
@@ -30,8 +30,8 @@ use std::{collections::HashMap, sync::Arc};
 pub struct Commitments<N: Network> {
     #[derivative(Debug = "ignore")]
     tree: Arc<MerkleTree<N::CommitmentsTreeParameters>>,
-    commitments: HashMap<N::Commitment, u32>,
-    current_index: u32,
+    commitments: HashMap<N::Commitment, u64>,
+    current_index: u64,
 }
 
 impl<N: Network> CommitmentsTreeScheme<N> for Commitments<N> {
@@ -47,9 +47,9 @@ impl<N: Network> CommitmentsTreeScheme<N> for Commitments<N> {
         })
     }
 
-    /// TODO (howardwu): Add safety checks for u32 (max 2^32).
+    /// TODO (howardwu): Add safety checks for u64 (max 2^42).
     /// Adds the given commitment to the tree, returning its index in the tree.
-    fn add(&mut self, commitment: &N::Commitment) -> Result<u32> {
+    fn add(&mut self, commitment: &N::Commitment) -> Result<u64> {
         // Ensure the commitment does not already exist in the tree.
         if self.contains_commitment(commitment) {
             return Err(MerkleError::Message(format!("{} already exists in the commitments tree", commitment)).into());
@@ -63,9 +63,9 @@ impl<N: Network> CommitmentsTreeScheme<N> for Commitments<N> {
         Ok(self.current_index - 1)
     }
 
-    /// TODO (howardwu): Add safety checks for u32 (max 2^32).
+    /// TODO (howardwu): Add safety checks for u64 (max 2^42).
     /// Adds all given commitments to the tree, returning the start and ending index in the tree.
-    fn add_all(&mut self, commitments: Vec<N::Commitment>) -> Result<(u32, u32)> {
+    fn add_all(&mut self, commitments: Vec<N::Commitment>) -> Result<(u64, u64)> {
         // Ensure the list of given commitments is non-empty.
         if commitments.is_empty() {
             return Err(anyhow!("The list of given commitments must be non-empty"));
@@ -91,10 +91,10 @@ impl<N: Network> CommitmentsTreeScheme<N> for Commitments<N> {
             commitments
                 .into_iter()
                 .enumerate()
-                .map(|(index, commitment)| (commitment, start_index + index as u32)),
+                .map(|(index, commitment)| (commitment, start_index + index as u64)),
         );
 
-        self.current_index += num_commitments as u32;
+        self.current_index += num_commitments as u64;
         let end_index = self.current_index - 1;
 
         Ok((start_index, end_index))
@@ -106,7 +106,7 @@ impl<N: Network> CommitmentsTreeScheme<N> for Commitments<N> {
     }
 
     /// Returns the index for the given commitment, if it exists.
-    fn get_commitment_index(&self, commitment: &N::Commitment) -> Option<&u32> {
+    fn get_commitment_index(&self, commitment: &N::Commitment) -> Option<&u64> {
         self.commitments.get(commitment)
     }
